fix: avoid passing project object to child components to prevent comlink serialization error
Use callback pattern (onIframeReady) instead of passing project object directly to Preview component.
This fixes the 'Cannot convert object to primitive value' error caused by comlink trying to serialize
the hmrServer property when the project proxy is accessed in child components.
